+++ src/main/java/net/silentchaos512/gear/init/ModItems.java
@@ -11,6 +11,7 @@ import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.parts.PartType;
 import net.silentchaos512.gear.config.Config;
 import net.silentchaos512.gear.item.*;
+import net.silentchaos512.gear.item.blueprint.book.BlueprintBookItem;
 import net.silentchaos512.gear.item.blueprint.GearBlueprintItem;
 import net.silentchaos512.gear.item.blueprint.PartBlueprintItem;
 import net.silentchaos512.gear.item.gear.*;
@@ -25,6 +26,8 @@ public final class ModItems {
             new BlueprintPackageItem(SilentGear.getId("starter_blueprints")));
 
     //region Blueprints and templates
+    public static final ItemRegistryObject<BlueprintBookItem> BLUEPRINT_BOOK = register("blueprint_book", () ->
+            new BlueprintBookItem(unstackableProps()));
     // Blueprints
     public static final ItemRegistryObject<PartBlueprintItem> ROD_BLUEPRINT = registerPartBlueprint(PartType.ROD, false);
     public static final ItemRegistryObject<PartBlueprintItem> TIP_BLUEPRINT = registerPartBlueprint(PartType.TIP, false);