+++ src/main/java/net/silentchaos512/gear/item/blueprint/PartBlueprintItem.java
@@ -6,6 +6,7 @@ import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.Tag;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
+import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.parts.PartType;
 
 public class PartBlueprintItem extends AbstractBlueprintItem {
@@ -20,6 +21,16 @@ public class PartBlueprintItem extends AbstractBlueprintItem {
 
     public PartType getPartType() {
         return partType;
+    }
+
+    @Override
+    public PartType getPartType(ItemStack stack) {
+        return partType;
+    }
+
+    @Override
+    public GearType getGearType(ItemStack stack) {
+        return GearType.PART;
     }
 
     @Override